重构线程管理逻辑，将 Thread 类重命名为 Worker，并更新相关实现；新增 Scheduler 类以支持多工作线程管理

+++ coruring/src/runtime/io/io_uring.cpp
@@ -51,7 +51,7 @@ void coruring::runtime::detail::IoUring::try_submit() {
 }
 
 coruring::runtime::detail::IoUring::IoUring(const Config& config)
-    : submit_interval_{config.submit_interval} {
+    : submit_interval_{static_cast<uint32_t>(config.submit_interval)} {
     if (auto ret = io_uring_queue_init(config.entries, &ring_, 0); ret < 0) [[unlikely]] {
         throw std::runtime_error(std::format("io_uring_queue_init failed: {}", strerror(-ret)));
     }
